+++ backend/funco-trade/src/main/java/com/found_404/funco/trade/domain/OpenTrade.java
@@ -41,13 +41,13 @@ public class OpenTrade extends BaseEntity {
     private Long orderCash;
 
     @Comment("가격")
-    private Long price;
+    private Double price;
 
     @Comment("구매한 가격")
-    private Long buyPrice;
+    private Double buyPrice;
 
     @Builder
-    public OpenTrade(Long memberId, String ticker, TradeType tradeType, Double volume, Long orderCash, Long price, Long buyPrice) {
+    public OpenTrade(Long memberId, String ticker, TradeType tradeType, Double volume, Long orderCash, Double price, Double buyPrice) {
         this.memberId = memberId;
         this.ticker = ticker;
         this.tradeType = tradeType;
@@ -57,7 +57,7 @@ public class OpenTrade extends BaseEntity {
         this.buyPrice = buyPrice;
     }
 
-    public static Trade toTrade(OpenTrade openTrade, Long tradePrice) {
+    public static Trade toTrade(OpenTrade openTrade, Double tradePrice) {
         return Trade.builder()
             .ticker(openTrade.getTicker())
             .volume(openTrade.getVolume())
