+++ backend/funco-trade/src/main/java/com/found_404/funco/trade/dto/response/MarketTradeResponse.java
@@ -14,6 +14,6 @@ public record MarketTradeResponse(
         Double volume,
         @NotNull
         @Positive
-        Long price
+        Double price
 ) {
 }
